Split `Params` from `format.rs`
