+++ CHANGELOG.adoc
@@ -16,6 +16,12 @@ project adheres to https://semver.org/[Semantic Versioning].
 
 toc::[]
 
+== {compare-url}/v0.6.0\...HEAD[Unreleased]
+
+=== Changed
+
+* Split `Params` from `format.rs` ({pull-request-url}/38[#38])
+
 == {compare-url}/v0.5.3\...v0.6.0[0.6.0] - 2023-08-09
 
 === Added
